clip-service/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=10.0.0
pika>=1.3.2
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
minio>=7.2.0
